feat: migratate docopt -> click
docopt-ng is not active project. the last release is 2023.
this pr migrate cli parameter handling library docopt-ng to click.
click is part of pallets ecosystem and it is very easy to define and
parse argument. I think, it is suitable for this project because
jageocoder cli include many subcommand.

This code was written by chatgpt codex.